move files out of dash folder (#1982)
