+++ app/assets/javascripts/dashboard/views/advanced_dataset/advanced_dataset_view.js
