+++ app/assets/javascripts/dashboard/views/advanced_dataset/dataset_access_list_view.js
