+++ app/assets/javascripts/dashboard/views/advanced_dataset/dataset_action_view.js
