+++ app/assets/javascripts/dashboard/views/dashboard_task_list_view.js
