+++ app/assets/javascripts/dashboard/views/dashboard_view.js
