+++ app/assets/javascripts/dashboard/views/dataset_view.js
