+++ app/assets/javascripts/dashboard/views/explorative_annotations_view.js
