+++ app/assets/javascripts/dashboard/views/logged_time_view.js
