+++ app/assets/javascripts/dashboard/views/spotlight_item_view.js
