+++ app/assets/javascripts/dashboard/views/transfer_task_modal.js
